refactor: structure post-inference passes by role
