+++ crates/semantics/src/checker/infer/expressions/functions.rs
@@ -182,7 +182,7 @@ impl TaskState<'_> {
         };
 
         // `Type::ignored()` defers the tail-position check to
-        // `validators/unused_expressions.rs`, which honors `#[allow(unused_*)]`.
+        // `passes/fact_producers/unused_expressions.rs`, which honors `#[allow(unused_*)]`.
         let has_implicit_unit_return = return_annotation == Annotation::Unknown;
         let body_ty = if has_implicit_unit_return {
             Type::ignored()
@@ -264,7 +264,7 @@ impl TaskState<'_> {
         // even when the closure is lexically inside a loop.
         let saved_loop_depth = self.scopes.reset_loop_depth();
         // `Type::ignored()` defers the tail-position check to
-        // `validators/unused_expressions.rs`, which honors `#[allow(unused_*)]`.
+        // `passes/fact_producers/unused_expressions.rs`, which honors `#[allow(unused_*)]`.
         let relax_body_to_unit = return_annotation == Annotation::Unknown && return_ty.is_unit();
         let body_ty = if relax_body_to_unit {
             Type::ignored()
