+++ crates/semantics/src/passes/checks/native_value_usage.rs
@@ -5,7 +5,7 @@ use syntax::types::{Symbol, Type, unqualified_name};
 
 use crate::store::Store;
 
-pub(super) fn run(typed_ast: &[Expression], module_id: &str, store: &Store, sink: &LocalSink) {
+pub(crate) fn run(typed_ast: &[Expression], module_id: &str, store: &Store, sink: &LocalSink) {
     for item in typed_ast {
         visit_expression(item, false, false, module_id, store, sink);
     }
