+++ crates/semantics/src/passes/checks/pattern_analysis/normalize.rs
