+++ crates/semantics/src/passes/checks/stringer_signature.rs
@@ -8,7 +8,7 @@ use diagnostics::LocalSink;
 use syntax::ast::Expression;
 use syntax::types::{SimpleKind, Type};
 
-pub(super) fn run(typed_ast: &[Expression], sink: &LocalSink) {
+pub(crate) fn run(typed_ast: &[Expression], sink: &LocalSink) {
     for item in typed_ast {
         visit(item, sink);
     }
