+++ crates/semantics/src/passes/checks/temp_producing.rs
@@ -2,7 +2,7 @@ use diagnostics::LocalSink;
 use syntax::ast::{Expression, FormatStringPart, Literal};
 use syntax::program::ReceiverCoercion;
 
-pub(super) fn run(typed_ast: &[Expression], sink: &LocalSink) {
+pub(crate) fn run(typed_ast: &[Expression], sink: &LocalSink) {
     for item in typed_ast {
         visit_expression(item, sink);
     }
