+++ crates/semantics/src/passes/lints/ast_walk/checks.rs
