+++ crates/semantics/src/passes/lints/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod ast_walk;
+pub(crate) mod from_facts;
+pub(crate) mod ref_graph;
+pub(crate) mod replaceable_with_zero_fill;
+
+pub use from_facts::Lint;
