+++ tests/_harness/lint.rs
@@ -1,5 +1,5 @@
 use diagnostics::{LisetteDiagnostic, LocalSink};
-use semantics::{checker::TaskState, store::Store, validators};
+use semantics::{checker::TaskState, passes, store::Store};
 use syntax::{
     desugar,
     lex::Lexer,
@@ -85,7 +85,7 @@ pub fn lint(source: &str) -> Vec<LisetteDiagnostic> {
 
     let analysis = semantics::context::AnalysisContext::new(&store, &checker.ufcs_methods);
     let mut unused = UnusedInfo::default();
-    validators::run(&analysis, &mut checker.facts, &sink, &mut unused, true);
+    passes::run(&analysis, &mut checker.facts, &sink, &mut unused, true);
 
     sink.take()
 }
